Remove cloudfront legacy redirect & associated DNS name for `train` environment only
This is being done as a quick fix.
The legacy dashboard CF distribution currently points to an account-level resource for well known environments.
This is fine for the most part, except the UAT account, which has 2 well known-environments in.
Which was causing a name clash when the `uat` and `train` envs try to create their CF distributions.