chore: increase parallelism on terraform apply invocations
We can save a bit of time by using increased parallelism with the
terraform apply and destroy commands. In local testing as part of
cdd-2118 this saved around 3 mins on average, so not loads but not
nothing. We could increase this further but that seems to be more
variable (at 30 I've had runs that saved another 30 seconds but also
runs that took an extra 4 mins over the base time with the default
parallelism, probably due to rate limiting?).